requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.21
click==8.1.7
blinker==1.6.2
itsdangerous==2.1.2
PyYAML